Fix RWLock tests to make them work on Windows
In the RWLock time tests, we need to consider situations where the
resolution of the system clock is quite small or, equivalently, 
the threads run really fast. 
In such cases there might be a situation when a reader exits 
the critical section and at the same time (according to our system
clock), a writer enters the critical section. Thus, we cannot assume
that the time of the former precedes the time of the latter (according
to the system clock of course).